Move GpuJoinUtils out of shims
