+++ sql-plugin/src/main/scala/org/apache/spark/sql/rapids/execution/GpuBroadcastNestedLoopJoinExec.scala
@@ -18,7 +18,7 @@ package org.apache.spark.sql.rapids.execution
 
 import ai.rapids.cudf.{ast, GatherMap, NvtxColor, OutOfBoundsPolicy, Table}
 import com.nvidia.spark.rapids._
-import com.nvidia.spark.rapids.shims.{GpuJoinUtils, ShimBinaryExecNode}
+import com.nvidia.spark.rapids.shims.ShimBinaryExecNode
 
 import org.apache.spark.TaskContext
 import org.apache.spark.broadcast.Broadcast
